Fix bug with request body
Found a bug when sending a POST request with a body. It appears that if
you don't pass the body in to http.NewRequest() then the Content-Length
header won't be properly set in the case where the body is a
bytes.Reader and then won't send the body at all. We could duplicate the
logic ourselves, but I figured if NewRequest() is already doing it, then
we can rearrange a bit to get the body reader first and then pass it
into NewRequest.